[tweakable] extract tweakable to its own module
move le_tweakable to its own module - this is much cleaner, and allows
us to treat it just like any other module (in terms of dependencies,
etc.)
