+++ modules/le_tweaks/le_tweaks.cpp
@@ -1,60 +1,29 @@
-#ifndef GUARD_PAL_TWEAKABLE_H
-#define GUARD_PAL_TWEAKABLE_H
+#include "le_tweaks.h"
+#include "le_core.h"
 
-/*
+#include <stdint.h>
+#include "le_core.h"
 
-  Tweakable allows you to tweak numerical parameters. It will only work in debug mode,
-  and in Release mode it will melt down to nothing.
+#include "le_file_watcher.h"
+#include "le_log.h" ///< if you get an error message saying that this header can't be found, make sure that the module that uses le_tweakable has a line saying `depends_on_island_module(le_log)` in its CMake file.
 
-  Each compilation unit which has tweaks must include this header, and must, at the
-  most convenient time, call
+#include <fstream>
+#include <iostream>
 
-        UPDATE_TWEAKS();
-
-  Which is the polling method for tweaks. Calling this will trigger callbacks if source
-  file changes have been detected via the file watcher.
-
-  To tweak individual values, set them as such:
-
-        int myVal = TWEAK(10);
-
-  Important: You must only place one tweakable value per line.
-
-  Note: This header file includes a few headers - which is a break with the #1 rule,
-  but there's no good way around it for now...
-
-  Thanks to Dennis Gustafsson, who originally described this technique in his blog:
-  http://blog.tuxedolabs.com/2018/03/13/hot-reloading-hardcoded-parameters.html
-
-*/
-
-#ifndef NDEBUG
-
-#	include <stdint.h>
-#	include "le_core.h"
-
-#	include "le_file_watcher.h"
-#	include "le_log.h" ///< if you get an error message saying that this header can't be found, make sure that the module that uses le_tweakable has a line saying `depends_on_island_module(le_log)` in its CMake file.
-
-#	include <fstream>
-#	include <iostream>
-
-#	include <cstring>
-#	include <string>
-
-// ----------------------------------------------------------------------
+#include <cstring>
+#include <string>
 
 // ----------------------------------------------------------------------
 // We use `class` here purely for RAAI, to ensure the destructor
 // gets called if the object gets deleted.
 class FileWatcher : NoCopy, NoMove {
-	le_file_watcher_o* self = le_file_watcher_api_i->le_file_watcher_i.create();
+    le_file_watcher_o* self = le_file_watcher_api_i->le_file_watcher_i.create();
 
   public:
-	FileWatcher() = default;
-	~FileWatcher() {
-		le_file_watcher_api_i->le_file_watcher_i.destroy( self );
-	}
+    FileWatcher() = default;
+    ~FileWatcher() {
+        le_file_watcher_api_i->le_file_watcher_i.destroy( self );
+    }
 
 	operator auto() {
 		return self;
@@ -66,54 +35,9 @@ class FileWatcher : NoCopy, NoMove {
 // automatically when the compilation unit which contains it gets unloaded.
 static FileWatcher aux_source_watcher{};
 
-struct CbData {
+using CbData = le_tweaks_api::CbData;
 
-	enum Type : uint32_t {
-		u64,
-		i64,
-		i32,
-		u32,
-		f32,
-		f64,
-		b32, // 32 bit bool
-	};
-
-	union Data {
-		uint64_t u64;
-		int64_t  i64;
-		double   f64;
-		uint32_t u32;
-		int32_t  i32;
-		float    f32;
-		bool     b32;
-		uint64_t raw;
-	};
-
-	uint32_t       line_num;
-	Type           type;
-	Data           data;
-	struct CbData* next; // linked list.
-
-#	define INITIALISER( T, TID )                          \
-		explicit CbData( uint32_t line_num_, TID param ) { \
-			line_num = line_num_;                          \
-			data.T   = param;                              \
-			type     = Type::T;                            \
-			next     = nullptr;                            \
-		}
-
-	INITIALISER( u64, uint64_t )
-	INITIALISER( u32, uint32_t )
-	INITIALISER( i32, int32_t )
-	INITIALISER( i64, int64_t )
-	INITIALISER( f32, float )
-	INITIALISER( f64, double )
-	INITIALISER( b32, bool )
-
-#	undef INITIALISER
-};
-
-static int tweakable_add_watch( CbData* cb_data, char const* file_path ) {
+static int le_tweaks_add_watch( CbData* cb_data, char const* file_path ) {
 
 	le_file_watcher_watch_settings watch;
 	watch.filePath           = file_path;
@@ -145,7 +69,7 @@ static int tweakable_add_watch( CbData* cb_data, char const* file_path ) {
 
 			if ( !file.is_open() ) {
 
-				LeLog( "le_tweakable" ).error( "Unable to open file: '%'", path );
+				LeLog( "le_tweaks" ).error( "Unable to open file: '%'", path );
 				return;
 			}
 
@@ -163,13 +87,13 @@ static int tweakable_add_watch( CbData* cb_data, char const* file_path ) {
 
 				char const* str_start = str.c_str();
 
-				static auto logger = LeLog( "le_tweakable" );
+				static auto logger = LeLog( "le_tweaks" );
 
 				while ( current_line_num == cb_data->line_num ) {
 
 					// Find chunk which actually designates data:
 
-					str_start = strstr( str_start, "TWEAK" );
+					str_start = strstr( str_start, "LE_TWEAK" );
 
 					if ( str_start == nullptr ) {
 						logger.warn( "Could not tweak line %d.", cb_data->line_num );
@@ -186,26 +110,26 @@ static int tweakable_add_watch( CbData* cb_data, char const* file_path ) {
 					//
 					switch ( cb_data->type ) {
 					case CbData::Type::u64:
-						sscanf( str_start, "TWEAK ( %lu ) ", &cb_data->data.u64 );
+						sscanf( str_start, "LE_TWEAK ( %lu ) ", &cb_data->data.u64 );
 						break;
 					case CbData::Type::i64:
-						sscanf( str_start, "TWEAK ( %li ) ", &cb_data->data.i64 );
+						sscanf( str_start, "LE_TWEAK ( %li ) ", &cb_data->data.i64 );
 						break;
 					case CbData::Type::i32:
-						sscanf( str_start, "TWEAK ( %d ) ", &cb_data->data.i32 );
+						sscanf( str_start, "LE_TWEAK ( %d ) ", &cb_data->data.i32 );
 						break;
 					case CbData::Type::u32:
-						sscanf( str_start, "TWEAK ( %ud ) ", &cb_data->data.u32 );
+						sscanf( str_start, "LE_TWEAK ( %ud ) ", &cb_data->data.u32 );
 						break;
 					case CbData::Type::f32:
-						sscanf( str_start, "TWEAK ( %f ) ", &cb_data->data.f32 );
+						sscanf( str_start, "LE_TWEAK ( %f ) ", &cb_data->data.f32 );
 						break;
 					case CbData::Type::f64:
-						sscanf( str_start, "TWEAK ( %lf ) ", &cb_data->data.f64 );
+						sscanf( str_start, "LE_TWEAK ( %lf ) ", &cb_data->data.f64 );
 						break;
 					case CbData::Type::b32: {
 						char token[ 6 ];
-						sscanf( str_start, "TWEAK ( %5c ) ", token );
+						sscanf( str_start, "LE_TWEAK ( %5c ) ", token );
 						if ( strncmp( token, "true", 4 ) == 0 ) {
 							cb_data->data.b32 = true;
 						} else if ( strncmp( token, "false", 5 ) == 0 ) {
@@ -221,7 +145,7 @@ static int tweakable_add_watch( CbData* cb_data, char const* file_path ) {
 						// Applied tweak.
 						long long   len = strstr( str_start, ")" ) - str_start;
 						std::string s   = { str_start, str_start + len + 1 };
-						logger.info( "Applied tweak at line #", current_line_num );
+						logger.info( "Applied tweak at line #%d", current_line_num );
 					}
 
 					// -- Check if there is a next tweak in our linked list of tweaks.
@@ -231,7 +155,7 @@ static int tweakable_add_watch( CbData* cb_data, char const* file_path ) {
 					if ( cb_data->next ) {
 						if ( cb_data->next->line_num == cb_data->line_num ) {
 							// We offset `str_start` by one, so that we can begin searching
-							// for the next `TWEAK` token on the same line.
+							// for the next `LE_TWEAK` token on the same line.
 							cb_data = cb_data->next;
 							str_start++;
 							continue;
@@ -264,31 +188,17 @@ static int tweakable_add_watch( CbData* cb_data, char const* file_path ) {
 	}
 };
 
-	// ----------------------------------------------------------------------
+// ----------------------------------------------------------------------
 
-#	define TWEAK( x )                                                                              \
-		[]( auto val, uint32_t line, char const* file_path )                                        \
-		    -> decltype( val )& {                                                                   \
-			static CbData cb_data( line, val );                                                     \
-			static int    val_watch = tweakable_add_watch( &cb_data, file_path );                   \
-			( void )val_watch; /* <- this does nothing, only to suppress unused variable warning */ \
-			return reinterpret_cast<decltype( val )&>( cb_data.data );                              \
-		}( x, __LINE__, __FILE__ )
+static void le_tweaks_update() {
+	le_file_watcher_api_i->le_file_watcher_i.poll_notifications( aux_source_watcher );
+}
 
-	// ----------------------------------------------------------------------
+// ----------------------------------------------------------------------
 
-#	define UPDATE_TWEAKS() \
-		le_file_watcher_api_i->le_file_watcher_i.poll_notifications( aux_source_watcher )
+LE_MODULE_REGISTER_IMPL( le_tweaks, api ) {
+	auto& le_tweaks_i = static_cast<le_tweaks_api*>( api )->le_tweaks_i;
 
-#else
-
-// In Release, we don't want to be able to tweak - therefore all tweak macros evaluate to
-// the values themselves.
-#	define TWEAK( x ) x
-
-// There is nothing to update, therefore this macro is a no-op.
-#	define UPDATE_TWEAKS()
-
-#endif // ifndef NDEBUG
-
-#endif // GUARD_PAL_TWEAKABLE_H
+	le_tweaks_i.update    = le_tweaks_update;
+	le_tweaks_i.add_watch = le_tweaks_add_watch;
+}
